backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==39.0.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
fastjsonschema==2.17.1
flatbuffers==23.5.26
flit_core==3.9.0
fonttools==4.40.0
frozenlist==1.3.3
gast==0.4.0
google==3.0.0
google-auth==2.20.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlesearch-python==1.2.3
grpcio==1.48.1
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.1
jupyter-events==0.6.3
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.10.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
multidict==6.0.4
munkres==1.1.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==8.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==3.6.0
pooch==1.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
# protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pyOpenSSL==23.2.0
pyparsing==3.1.0
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pyu2f==0.1.5
# windows specific, and thus not suited for docker, but locally required
# pywin32==304
# pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==67.7.2
sip==6.5.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
termcolor==2.3.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
typing-utils==0.1.0
tzdata==2023.3
unicodedata2==15.0.0
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.6
wheel==0.40.0
widgetsnbextension==4.0.7
wikipedia==1.4.0
win-inet-pton==1.1.0
wrapt==1.15.0
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
